requirements.txt
amqp==2.6.1
appnope==0.1.2
asgiref==3.2.10
autopep8==1.5.3
backcall==0.2.0
billiard==3.6.4.0
bleach==3.3.0
CacheControl==0.12.6
cachetools==4.2.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.9
dj-database-url==0.5.0
Django==3.0.8
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-extensions==3.0.9
django-markdownify==0.9.0
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
firebase-admin==4.5.2
flower==0.9.7
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==20.0.4
httplib2==0.19.0
humanize==3.6.0
idna==2.10
inflection==0.5.1
ipython==7.24.1
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
kombu==4.6.11
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
msgpack==1.0.2
mypy==0.812
mypy-extensions==0.4.3
natsort==7.1.1
nose==1.3.7
numpy==1.20.3
openapi-codec==1.3.2
packaging==20.8
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.18
proto-plus==1.15.0
protobuf==3.15.6
psycopg2==2.8.5
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.9.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.17.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
searchtweets==1.7.6
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
tornado==6.1
traitlets==5.0.5
tweet-parser==1.13.2
twilio==6.59.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
unicodecsv==0.14.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.1.0
